Fix for Ember 3.10
This change moves us away from depending on implementation details of computed properties. Rather than using type information (`ComputedProperty`), and accessing the private `_dependentKeys` and `_getter` properties off of `ComputedProperty` objects (which aren't all exposed in Ember 3.10), we store off the needed information in a weak map so we can access it directly without depending on the internals of Ember's computed property implementation.